Refactor calendar components for improved organization and maintainability
- Consolidated all calendar-related components into organized subfolders for better structure.
- Updated import statements in CalendarScreen to reflect the new component organization.
- Ensured backward compatibility by re-exporting components from the main CalendarComponents file.
- Enhanced code readability and maintainability by streamlining the component structure.
